Move setHeaders function to ob-util.js
